+++ .gitignore
@@ -8,5 +8,3 @@ yarn.lock
 # Build directory
 /public
 .DS_Store
-
-interludio-*.json
