+++ package.json
@@ -18,6 +18,7 @@
     "gatsby-plugin-offline": "^2.2.10",
     "gatsby-plugin-react-helmet": "^3.1.13",
     "gatsby-plugin-sass": "^2.1.20",
+    "gatsby-source-graphql": "^2.5.1",
     "node-sass": "^4.12.0",
     "react": "^16.10.2",
     "react-burger-menu": "^2.6.11",
