+++ src/pages/mi-historia.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import { Link } from 'gatsby';
 
 import Layout from '../components/layout';
 import Footer from '../components/Footer';
 import Header from '../components/Header';
+import Testimonial from '../components/Testimonial';
 
-const SecondPage = () => (
+const MiHistoria = ({ data }) => (
   <Layout>
     <Header />
 
@@ -12,35 +14,14 @@ const SecondPage = () => (
       <section id="main" className="wrapper">
         <div className="inner">
           <h2>Celebramos 12 años juntos</h2>
-          <p>Contános, ¿que es Interludio para vos?</p>
-          <div className="style1">
-            <section>
-              <form
-                name="testimonials"
-                method="post"
-                action="/interludio-para-mi"
-                data-netlify="true"
-                data-netlify-recaptcha="true"
-              >
-                <input type="hidden" name="bot-field" />
-                <input type="hidden" name="form-name" value="contact" />
-                <div className="fields">
-                  <div className="field">
-                    <label htmlFor="mensaje">Mensaje</label>
-                    <textarea name="mensaje" id="message" rows="5" />
-                  </div>
-                  <div className="field half">
-                    <label htmlFor="name">Nombre</label>
-                    <input type="text" name="nombre" id="name" />
-                  </div>
-                </div>
-                <ul className="actions">
-                  <button type="submit" className="button submit">
-                    Guardar
-                  </button>
-                </ul>
-              </form>
-            </section>
+          <div>¿Qué es Interludio para vos?</div>
+          <Link className="button" to="/contanos-tu-historia">
+            Contanos
+          </Link>
+          <div className="testimonials">
+            {data.hasura.testimonials.map(test => (
+              <Testimonial content={test.comment} name={test.name} />
+            ))}
           </div>
         </div>
       </section>
@@ -50,4 +31,15 @@ const SecondPage = () => (
   </Layout>
 );
 
-export default SecondPage;
+export default MiHistoria;
+
+export const query = graphql`
+  query Testimonials {
+    hasura {
+      testimonials {
+        name
+        comment
+      }
+    }
+  }
+`;
